+++ next-docs/public/examples/circularProgressTw/Values.tsx
@@ -1,0 +1,12 @@
+import { CircularProgress } from '@heathmont/moon-core-tw';
+
+const Example = () => (
+  <>
+    <CircularProgress value={25} />
+    <CircularProgress value={50} />
+    <CircularProgress value={75} />
+    <CircularProgress value={100} />
+  </>
+);
+
+export default Example;