+++ packages/sidebar/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 9.11.2
+
+### Patch Changes
+
+- Add Lab and Partners themes
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.11.2
+  - @heathmont/moon-utils@9.11.2
+
 ## 9.11.1
 
 ### Patch Changes